+++ @xen-orchestra/web/src/pages/host/[id]/dashboard.vue
@@ -1,5 +1,5 @@
 <template>
-  <div class="dashboard" :class="{ mobile: isMobile }">
+  <div class="dashboard" :class="{ mobile: uiStore.isMobile }">
     <HostDashboardQuickInfo class="quick-info" :host />
     <div v-if="!isHostRunning" class="offline-hero-container">
       <VtsStateHero format="page" type="offline" size="large" horizontal>
@@ -66,7 +66,7 @@ const { hostAlarms, areHostAlarmsReady, hasHostAlarmFetchError } = useXoHostAlar
 
 const isHostRunning = computed(() => host.power_state === HOST_POWER_STATE.RUNNING)
 
-const { isMobile } = useUiStore()
+const uiStore = useUiStore()
 </script>
 
 <style lang="postcss" scoped>
@@ -83,7 +83,7 @@ const { isMobile } = useUiStore()
     'offline-hero-container offline-hero-container offline-hero-container offline-hero-container offline-hero-container offline-hero-container offline-hero-container offline-hero-container';
 
   &.mobile {
-    grid-template-columns: 1fr;
+    grid-template-columns: minmax(0, 1fr);
     grid-template-areas:
       'quick-info'
       'alarms'